requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
api==0.0.7
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
boto3==1.26.123
botocore==1.29.123
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.3.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
nose==1.3.7
numpy==1.24.2
openai==0.27.5
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycountry==22.3.5
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.2
requests==2.29.0
rich==13.3.5
s3transfer==0.6.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
